+++ CHANGELOG.md
@@ -50,6 +50,7 @@ at anytime.
   * integration tests for bootstrapping the dht
   * configurable `concurrent_announcers` setting
   * `peer_ping` command
+  * `--sort` option in `file_list`
 
 ### Removed
   * `announce_all` argument from `blob_announce`